Restructure status stop
